Add test, move func to utilities, fix whoops
